feat: add contact labels feature with shared LabelActions component
- Extract common label management logic into generic LabelActions component
- Refactor ConversationLabelActions to use shared component
- Create ContactLabelActions wrapper for contact label management
- Integrate ContactLabelActions into ContactDetailsScreen
- Fix contact labels API method from PUT to POST to match API spec
