+++ src/components-next/label-section/index.ts
@@ -1,2 +1,3 @@
 export * from './LabelCell';
 export * from './LabelItem';
+export * from './LabelActions';
