+++ src/screens/contact-details/components/index.ts
@@ -1,3 +1,4 @@
 export * from './ContactDetailsScreenHeader';
 export * from './ContactBasicActions';
 export * from './ContactMetaInformation';
+export * from './ContactLabelActions';
